Store length and capacity of Payloads object as u32 (#130)
Currentlly, `wasmtime_continuations::Payloads` looks like this (after
inlining some definitions):

```rust
pub struct Payloads {
    pub length: usize,
    pub capacity: usize,
    pub data: *mut u128,
}
```

This PR changes the type of the `capacity` and `length` fields from
`usize` to `u32`. This is reasonable, as the maximum length/capacity
values we will see are bounded by the maximum of the following:
- Number of parameters/return values of any function used as a
continuation
- Number of parameters/return values of any tag

Of course, it would be nice to somehow enforce the actual maximum of
these values that we support, but that seems out of scope for this PR.

Each `ContinuationObject` contains two such `Payloads` objects, so this
change saves us 16 bytes per continuation. This is probably not a
meaningful optimization by itself, but shrinking these values is useful
for subsequent work where it may come in handy when I'm packing parts of
this in an enum in a custom way.

This PR also adds a `repr(C)` to `Payloads`, which should have been
there the whole time, but wasn't.

+++ crates/continuations/src/lib.rs
@@ -22,9 +22,9 @@ pub mod types {
     /// Types used by `Payloads` struct
     pub mod payloads {
         /// type of length
-        pub type Length = usize;
+        pub type Length = u32;
         /// type of capacity
-        pub type Capacity = usize;
+        pub type Capacity = u32;
         /// Type of the entries in the actual buffer
         pub type DataEntries = u128;
     }
@@ -76,6 +76,7 @@ impl StackLimits {
 unsafe impl Send for StackLimits {}
 unsafe impl Sync for StackLimits {}
 
+#[repr(C)]
 pub struct Payloads {
     /// Number of currently occupied slots.
     pub length: types::payloads::Length,
@@ -87,11 +88,11 @@ pub struct Payloads {
 }
 
 impl Payloads {
-    pub fn new(capacity: usize) -> Self {
+    pub fn new(capacity: u32) -> Self {
         let data = if capacity == 0 {
             ptr::null_mut()
         } else {
-            let mut args = Vec::with_capacity(capacity);
+            let mut args = Vec::with_capacity(capacity as usize);
             let args_ptr = args.as_mut_ptr();
             args.leak();
             args_ptr
