+++ crates/environ/src/builtin.rs
@@ -40,7 +40,7 @@ macro_rules! foreach_builtin_function {
             new_epoch(vmctx: vmctx) -> i64;
 
             /// Creates a new continuation from a funcref.
-            tc_cont_new(vmctx: vmctx, r: pointer, param_count: i64, result_count: i64) -> pointer;
+            tc_cont_new(vmctx: vmctx, r: pointer, param_count: i32, result_count: i32) -> pointer;
             /// Resumes a continuation. The result value is of type
             /// wasmtime_continuations::SwitchDirection.
             tc_resume(vmctx: vmctx, contobj: pointer, parent_stack_limits: pointer) -> i64;
